requirements.txt
Django
django-cors-headers
django-filter
django-rest-swagger
djangorestframework~=3.14.0
django-environ
drf-yasg
Pillow
psycopg2
# django-redis
